python/pyproject.toml
[project]
name = "aigov-py"
version = "0.2.1"
description = "AIGov Python package: thin govai audit HTTP client (import govai) and related tooling."
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "scikit-learn>=1.4.0",
  "joblib>=1.3.0",
  "PyYAML>=6.0.0",
  "PyNaCl>=1.5.0",
  "sigstore>=4.2.0"
]

[project.scripts]
govai = "aigov_py.cli:main"

[project.optional-dependencies]
server = [
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0"
]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "mypy>=1.8.0",
  "types-requests>=2.31.0.20240106"
]
runs-postgres = [
  "psycopg[binary]>=3.2.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
aigov_py = [
  "experiments/datasets/*.json",
  "experiments/scripts/*.py",
  "experiments/scenario_rubric.json",
  "experiments/artifact_bundle_replay_rubric.json",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["aigov_py*", "govai*"]
exclude = ["artifacts*", "tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
